Use self-attention with keras MultiheadAttention layer.
Before, I was guessing as to the necessity for v/k/q, and it appears
that self-attention does away with this altogether. I'm not sure why
yet, but "self-attention" will be something I look up in the future if
this ends up working. I probably will anyways.

I also made a change to the way the positional encodings are applied.
Before, I applied the positional encodings before a NN leading into the
attention mechanism. Now I'm applying it AFTER the NN and before the
attention mechanism. This is also closer to the initial implementation
in the original repo that I forked.